ssh: Hack to allow ssh agent forwarding to work
For machines that don't have their own keys we need to disable a few
things so they can use the forwarded agent.  One of those settings is
`IdentitiesOnly`.

This is currently based on the `keysDir` setting.  I might remove this
setting in the future and start moving all my keys back into ~/.ssh.
If I do that I should probably think about whether or not I need
`IdentitiesOnly` anymore.