laptop: Eventually hibernate
Prior to this commit I was solely using `suspend`.  After my laptop
sat around for a few days and the battery died I would have to boot
from scratch.

These changes make it so that either the hibernate data is written
when suspending or the laptop wakes up after an hour and hibernates.
That way there is always hibernate data to restart from.