atuin: Remove
There's a lot to like about Atuin.  However, I could never get it to
integrate with my normal shell (zsh) and the nix shell (bash).  And
the UI is a bit distracting.  I'd rather just find a fuzzy history
search that was built into zsh that doesn't take over part of the
screen.  I prefer the way zsh deals with completions.